feat(11107): Updated caches and data adapters pages
