+++ graylog2-web-interface/src/routing/AppRouter.tsx
@@ -257,34 +257,10 @@ const AppRouter = () => {
             !isCloud && { path: RoutePaths.SYSTEM.INDICES.FAILURES, element: <IndexerFailuresPage /> },
 
             { path: RoutePaths.SYSTEM.LOOKUPTABLES.OVERVIEW, element: <LUTTablesPage /> },
-            // { path: RoutePaths.SYSTEM.LOOKUPTABLES.CREATE, element: <LUTTablesPage action="create" /> },
-            // { path: RoutePaths.SYSTEM.LOOKUPTABLES.show(':tableName'), element: <LUTTablesPage action="show" /> },
-            // { path: RoutePaths.SYSTEM.LOOKUPTABLES.edit(':tableName'), element: <LUTTablesPage action="edit" /> },
 
             { path: RoutePaths.SYSTEM.LOOKUPTABLES.CACHES.OVERVIEW, element: <LUTCachesPage /> },
-            { path: RoutePaths.SYSTEM.LOOKUPTABLES.CACHES.CREATE, element: <LUTCachesPage action="create" /> },
-            {
-              path: RoutePaths.SYSTEM.LOOKUPTABLES.CACHES.show(':cacheName'),
-              element: <LUTCachesPage action="show" />,
-            },
-            {
-              path: RoutePaths.SYSTEM.LOOKUPTABLES.CACHES.edit(':cacheName'),
-              element: <LUTCachesPage action="edit" />,
-            },
 
             { path: RoutePaths.SYSTEM.LOOKUPTABLES.DATA_ADAPTERS.OVERVIEW, element: <LUTDataAdaptersPage /> },
-            {
-              path: RoutePaths.SYSTEM.LOOKUPTABLES.DATA_ADAPTERS.CREATE,
-              element: <LUTDataAdaptersPage action="create" />,
-            },
-            {
-              path: RoutePaths.SYSTEM.LOOKUPTABLES.DATA_ADAPTERS.show(':adapterName'),
-              element: <LUTDataAdaptersPage action="show" />,
-            },
-            {
-              path: RoutePaths.SYSTEM.LOOKUPTABLES.DATA_ADAPTERS.edit(':adapterName'),
-              element: <LUTDataAdaptersPage action="edit" />,
-            },
 
             { path: RoutePaths.SYSTEM.PIPELINES.OVERVIEW, element: <PipelinesOverviewPage /> },
             { path: RoutePaths.SYSTEM.PIPELINES.RULES, element: <RulesPage /> },
